+++ apps/action-config.js
@@ -62,6 +62,15 @@ export class ActionConfig extends FormApplication {
         });
     }
 
+    _activateFilePicker(event) {
+        event.preventDefault();
+        const options = this._getFilePickerOptions(event);
+        options.wildcard = true;
+        const fp = new FilePicker(options);
+        this.filepickers.push(fp);
+        return fp.browse();
+    }
+
     fillList(list, id) {
         return (list instanceof Array
             ? list.map(g => { return $('<optgroup>').attr('label', i18n(g.text)).append(Object.entries(g.groups).map(([k, v]) => { return $('<option>').attr('value', (g.id ? g.id + ":" : '') + k).html(i18n(v)).prop('selected', ((g.id ? g.id + ":" : '') + k) == id) })) })